backend/summary-keyword/requirements.txt
psycopg2
boto3
google-generativeai
pika